Move designate gm or player
Moved the player designation from the registration page to the create campaign page
