+++ campaignJournal/server/models/Campaign.js
@@ -4,6 +4,14 @@ module.exports = (sequelize, DataTypes) => {
             type: DataTypes.STRING,
             allowNull: false,
         },
+        gameMaster: {
+            type: DataTypes.INTEGER,
+            allowNull: true,
+            references: {
+                model: "Users",
+                key: "id",
+            }
+        }
     });
 
     return Campaign;
